+++ src/rpc/repl_session.h
@@ -53,11 +53,9 @@ namespace barch {
                         name.resize(name_size);
                         readp(stream, name.data(), name_size);
                         auto ks = get_keyspace(name);
-                        auto t = ks->get(shard);
                         ostream.clear();
                         // self calls will deadlock here
-                        write_lock release(t->get_latch());
-                        process_art_fun_cmd(t, name_size, ostream, stream.buf);
+                        process_art_fun_cmd(ks, name_size, ostream, stream.buf);
                         do_write(ostream);
                     }else {
                         // clients can disconnect for all sorts of reasons