+++ RELEASES.md
@@ -1,6 +1,8 @@
 # Release v0.4.3.0b 2025-12-23
 
 ## New Features
+1. make replication shard count agnostic
+2. add debug replication mode
 
 # Release v0.4.2.6b 2025-12-23